Cleanup

+++ Embraced/RCFTTaskViewController.swift
@@ -13,13 +13,6 @@ import FirebaseStorage
 class RCFTTaskViewController: ActiveStepViewController {
 
     let TEST_NAME = "RCFT"
-    var lastPoint = CGPoint.zero
-    var red: CGFloat = 0.0
-    var green: CGFloat = 0.0
-    var blue: CGFloat = 0.0
-    var brushWidth: CGFloat = 3.0
-    var opacity: CGFloat = 1.0
-    var swiped = false
     var documentPath: URL?
     let fileName = "rcft.png"
     var saveImagePath: URL!
@@ -34,12 +27,14 @@ class RCFTTaskViewController: ActiveStepViewController {
         return imageView
     }()
     
-    let canvas: UIImageView = {
-        var imageView = UIImageView()
-        imageView.translatesAutoresizingMaskIntoConstraints = false
-        imageView.layer.borderColor = UIColor.black.cgColor
-        imageView.layer.borderWidth = 0.5
-        return imageView
+    let canvas: CanvasView = {
+        var canvasView = CanvasView()
+        canvasView.translatesAutoresizingMaskIntoConstraints = false
+        canvasView.layer.borderColor = UIColor.black.cgColor
+        canvasView.layer.borderWidth = 0.5
+        canvasView.brushWidth = 3.0
+        canvasView.isUserInteractionEnabled = true
+        return canvasView
     }()
     
     override func viewDidLoad() {
@@ -102,79 +97,32 @@ class RCFTTaskViewController: ActiveStepViewController {
             }
         }
     }
-
-    override func touchesBegan(_ touches: Set<UITouch>, with event: UIEvent?) {
-        swiped = false
-        if let touch = touches.first {
-            lastPoint = touch.location(in: self.canvas)
-        }
-    }
-    
-    override func touchesMoved(_ touches: Set<UITouch>, with event: UIEvent?) {
-        swiped = true
-        if let touch = touches.first {
-            let currentPoint = touch.location(in: canvas)
-            drawLine(from: lastPoint, to: currentPoint)
-            
-            lastPoint = currentPoint
-        }
-    }
-    
-    override func touchesEnded(_ touches: Set<UITouch>, with event: UIEvent?) {
-        if !swiped {
-            self.drawLine(from: lastPoint, to: lastPoint)
-        }
-        nextBtn.isHidden = false
-    }
-    
-    
-    private func drawLine(from fromPoint: CGPoint, to toPoint: CGPoint) {
-        UIGraphicsBeginImageContextWithOptions(canvas.bounds.size, false, 0)
-        
-        canvas.image?.draw(in: canvas.bounds)
-        
-        let context = UIGraphicsGetCurrentContext()
-        
-        context?.move(to: fromPoint)
-        context?.addLine(to: toPoint)
-        
-        context?.setLineCap(CGLineCap.round)
-        context?.setLineWidth(brushWidth)
-        context?.setStrokeColor(red: red, green: green, blue: blue, alpha: 1.0)
-        context?.setBlendMode(CGBlendMode.normal)
-        context?.strokePath()
-        
-        canvas.image = UIGraphicsGetImageFromCurrentImageContext()
-        canvas.alpha = opacity
-        UIGraphicsEndImageContext()
-    }
     
     private func externalStorage() {
         SVProgressHUD.show()
         let filePath = "\(FirebaseStorageManager.shared.pid!)/\(TEST_NAME)/\(fileName)"
         
         if Utility.fileExist(filePath) {
-            let storage = Storage.storage()
-            let storageRef = storage.reference()
-            let participantRef = storageRef.child(filePath)
-            
-            participantRef.putFile(from: saveImagePath, metadata: nil) { (metadata, error) in
+            FirebaseStorageManager.shared.externalStorage(filePath: filePath, fileUrl: saveImagePath) { (uploadTask, error) in
                 if error != nil {
-                    print("Error: \(error?.localizedDescription)")
-                    SVProgressHUD.showError(withStatus: "An error has happened")
+                    SVProgressHUD.showError(withStatus: error?.localizedDescription)
                 }
                 
-                RCFTModel.shared.file_1 = filePath
-                RCFTModel.shared.time_1 = self.reactionTime
-                
-                FirebaseStorageManager.shared.addDataToDocument(payload: [
-                    "rcft": RCFTModel.shared.printModel()
-                ])
-                
-                self.performSegue(withIdentifier: "moveToDone", sender: nil)
-                
-                Utility.deleteFile(filePath)
-                SVProgressHUD.dismiss()
+                if let task = uploadTask {
+                    task.observe(.success, handler: { (snapshot) in
+                        RCFTModel.shared.file_1 = filePath
+                        RCFTModel.shared.time_1 = self.reactionTime
+                        
+                        FirebaseStorageManager.shared.addDataToDocument(payload: [
+                            "rcft": RCFTModel.shared.printModel()
+                        ])
+                        
+                        Utility.deleteFile(filePath)
+                        SVProgressHUD.dismiss()
+                        
+                        self.performSegue(withIdentifier: "moveToDone", sender: nil)
+                    })
+                }
             }
         }
     }
